fixed apply button hanler for TaskResultInput, moved back buttons to component
