+++ its_ext/src/components/BackButton/index.ts
@@ -1,0 +1,3 @@
+import { BackButton } from "./BackButton";
+
+export { BackButton };
