+++ its_ext/src/components/BackButton/types.ts
@@ -1,0 +1,3 @@
+export interface IBackButtonProps {
+  route: string;
+}
