+++ its_ext/src/components/StudentsAdmission/Base/StudentsAdmission.tsx
@@ -8,13 +8,10 @@ import { ICompetitionGroupItem } from "../CompetitionGroupSelect/types";
 import { REQUEST_ERROR_UNAUTHORIZED } from "../../../utils/constants";
 import { TaskResultsInput } from "../TaskResultsInput";
 
-import { useNavigate } from "react-router-dom";
-
-import { Button } from "@mui/material";
-import WestIcon from "@mui/icons-material/West";
 import { ApplyButtonWithActionDisplay } from "../../ApplyButtonWithActionDisplay";
 import { StudentsDistribution } from "../StudentsDistribution";
 import { SubgroupDistribution } from "../SubgroupDistribution";
+import { BackButton } from "../../BackButton";
 
 export function StudentsAdmission(props: IStudentsAdmissionProps) {
   const [competitionGroupItems, setCompetitionGroupItems] = useState<
@@ -27,11 +24,7 @@ export function StudentsAdmission(props: IStudentsAdmissionProps) {
   const stepThreeRef = useRef<HTMLElement | null>(null);
   const stepFourRef = useRef<HTMLElement | null>(null);
 
-  const navigate = useNavigate();
 
-  const handleBackButton = () => {
-    navigate("/");
-  };
 
   const isGroupSelectionValid = () => {
     return competitionGroupIds.length > 0 && competitionGroupIds.length <= 2;
@@ -149,19 +142,7 @@ export function StudentsAdmission(props: IStudentsAdmissionProps) {
   return (
     <section className="page">
       <h2 className="action_header">
-        <Button
-          onClick={handleBackButton}
-          variant="text"
-          style={{
-            position: "absolute",
-            left: 0,
-            top: "50%",
-            transform: "translate(0, -50%)",
-          }}
-          startIcon={<WestIcon />}
-        >
-          Вернуться назад
-        </Button>
+        <BackButton route="/" />
         Зачисление студентов
       </h2>
       <article className="step">
