+++ its_ext/src/index.css
@@ -7,6 +7,7 @@ body {
   -moz-osx-font-smoothing: grayscale;
 
   background-color: #c6c6c6;
+  /* background-color: #dbdbdb; */
 }
 
 code {
@@ -173,6 +174,7 @@ td {
 .table__сontainer {
   max-height: 600px;
   overflow: auto;
+  margin: 1em 0;
 }
 
 .table_vertical_borders td {
